+++ bugsnag-plugin-android-exitinfo/src/main/java/com/bugsnag/android/ApplicationExitInfoMatcher.kt
@@ -1,27 +1,18 @@
 package com.bugsnag.android
 
-import android.app.ActivityManager
 import android.app.ApplicationExitInfo
-import android.content.Context
 import android.os.Build
 import androidx.annotation.RequiresApi
 
 @RequiresApi(Build.VERSION_CODES.R)
 internal class ApplicationExitInfoMatcher(
-    private val context: Context,
-    private val pid: Int
+    private val applicationExitInfo: List<ApplicationExitInfo>,
+    private val previousState: ExitInfoPluginStore.PersistentState? = null,
 ) {
     fun matchExitInfo(event: Event): ApplicationExitInfo? {
-        val am: ActivityManager =
-            context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
-        val allExitInfo: List<ApplicationExitInfo> =
-            am.getHistoricalProcessExitReasons(context.packageName, MATCH_ALL, MAX_EXIT_INFO)
-        val sessionIdBytes: ByteArray =
-            event.session?.id?.toByteArray() ?: return null
-        val exitInfo: ApplicationExitInfo =
-            findExitInfoBySessionId(allExitInfo, sessionIdBytes)
-                ?: findExitInfoByPid(allExitInfo) ?: return null
-        return exitInfo
+        val sessionIdBytes: ByteArray = event.session?.id?.toByteArray() ?: return null
+        return findExitInfoBySessionId(applicationExitInfo, sessionIdBytes)
+            ?: findExitInfoByPid(applicationExitInfo)
     }
 
     internal fun findExitInfoBySessionId(
@@ -32,7 +23,7 @@ internal class ApplicationExitInfoMatcher(
     }
 
     internal fun findExitInfoByPid(allExitInfo: List<ApplicationExitInfo>) =
-        allExitInfo.find { it.pid == pid }
+        allExitInfo.find { it.pid == previousState?.pid }
 
     internal companion object {
         const val MATCH_ALL = 0