+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * `memoryTrimLevel` and `lowMemory` will not be reported when they are not set (newer versions of Android do not set these values)
   [#2237](https://github.com/bugsnag/bugsnag-android/pull/2237)
 
+### Bug fixes
+
+* Handle edge-cases to reduce the changes of duplicate ANR reporting from `bugsnag-plugin-android-exitinfo` when `reportUnmatchedANR = true && disableProcessStateSummaryOverride = true`
+  [#2235](https://github.com/bugsnag/bugsnag-android/pull/2235)
+
 ## 6.16.0 (2025-07-31)
 
 ### Changes